+++ main.go
@@ -59,6 +59,9 @@ func main() {
 			Assets: assets,
 		},
 		BackgroundColour: &options.RGBA{R: 0, G: 0, B: 0, A: 0},
+		DragAndDrop: &options.DragAndDrop{
+			DisableWebViewDrop: true,
+		},
 		Bind: []interface{}{
 			svc,
 		},